# Conflicts:
#	light_ws2812_AVR/Makefile

+++ light_ws2812_Arduino/light_WS2812/examples/fade_rgb/fade_rgb.ino
@@ -34,7 +34,7 @@ void setup() {
 }
 
 void loop() {
-	byte i = 0;
+	int i = 0;
 	if (sign) {
 		intensity++;
 		if (intensity == 255)
